testing: moved kde unit tests to test_kde.py
